+++ example/android/app/src/main/kotlin/com/ashkan/app_bar_search_example/MainActivity.kt
@@ -1,6 +1,0 @@
-package com.ashkan.app_bar_search_example
-
-import io.flutter.embedding.android.FlutterActivity
-
-class MainActivity: FlutterActivity() {
-}