+++ example/macos/Flutter/GeneratedPluginRegistrant.swift
@@ -5,8 +5,6 @@
 import FlutterMacOS
 import Foundation
 
-import app_bar_search
 
 func RegisterGeneratedPlugins(registry: FlutterPluginRegistry) {
-  AppBarSearchPlugin.register(with: registry.registrar(forPlugin: "AppBarSearchPlugin"))
 }