+++ example/windows/flutter/generated_plugin_registrant.cc
@@ -6,9 +6,6 @@
 
 #include "generated_plugin_registrant.h"
 
-#include <app_bar_search/app_bar_search_plugin_c_api.h>
 
 void RegisterPlugins(flutter::PluginRegistry* registry) {
-  AppBarSearchPluginCApiRegisterWithRegistrar(
-      registry->GetRegistrarForPlugin("AppBarSearchPluginCApi"));
 }