+++ ios/Classes/AppBarSearchPlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface AppBarSearchPlugin : NSObject<FlutterPlugin>
-@end